requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.4.6
    # via aiohttp
aiohttp==3.11.13
    # via instructor
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   openai
attrs==23.2.0
    # via aiohttp
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.7
    # via typer
distro==1.9.0
    # via openai
docstring-parser==0.16
    # via instructor
docx2txt==0.8
    # via -r requirements.in
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.27.0
    # via
    #   langsmith
    #   openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
instructor==1.7.3
    # via -r requirements.in
jinja2==3.1.6
    # via instructor
jiter==0.8.2
    # via
    #   instructor
    #   openai
joblib==1.4.2
    # via scikit-learn
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain-core==0.3.42
    # via
    #   -r requirements.in
    #   langchain-text-splitters
langchain-text-splitters==0.3.6
    # via -r requirements.in
langsmith==0.1.147
    # via langchain-core
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
numpy==1.26.4
    # via
    #   -r requirements.in
    #   pandas
    #   scikit-learn
    #   scipy
openai==1.64.0
    # via instructor
orjson==3.10.15
    # via langsmith
packaging==24.1
    # via langchain-core
pandas==2.2.3
    # via -r requirements.in
pdf2image==1.16.0
    # via -r requirements.in
pillow==10.0.0
    # via
    #   -r requirements.in
    #   pdf2image
propcache==0.3.0
    # via
    #   aiohttp
    #   yarl
pydantic==2.8.2
    # via
    #   -r requirements.in
    #   instructor
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.20.1
    # via
    #   instructor
    #   pydantic
pygments==2.18.0
    # via rich
pypdf2==3.0.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2025.1
    # via pandas
pyyaml==6.0.1
    # via langchain-core
regex==2024.5.15
    # via tiktoken
requests==2.32.3
    # via
    #   -r requirements.in
    #   instructor
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
rich==13.7.1
    # via
    #   instructor
    #   typer
scikit-learn==1.6.1
    # via -r requirements.in
scipy==1.15.2
    # via
    #   -r requirements.in
    #   scikit-learn
shellingham==1.5.4
    # via typer
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
tenacity==9.0.0
    # via
    #   instructor
    #   langchain-core
threadpoolctl==3.5.0
    # via scikit-learn
tiktoken==0.9.0
    # via -r requirements.in
tqdm==4.67.1
    # via
    #   -r requirements.in
    #   openai
typer==0.12.3
    # via instructor
typing-extensions==4.12.2
    # via
    #   -r requirements.in
    #   anyio
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   typer
tzdata==2024.1
    # via pandas
urllib3==2.3.0
    # via requests
yarl==1.18.3
    # via aiohttp
